# ARROW-12533: [C++] Add random real distribution function