@cyb70289 2021-05-10
**NOTE:** Below benchmark results are the same, but the total run time differs much. The saved time is from test preparation (generating random bools by calling bernoulli_distribution), not the test itself.

Test on Neoverse N1, clang10.

Before: **59.9s** total run time
```
$ time release/arrow-compute-aggregate-benchmark --benchmark_filter="ModeKernelNarrow<BooleanType>/1048576/10000"
------------------------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------
ModeKernelNarrow<BooleanType>/1048576/10000        637 us          637 us         1097 bytes_per_second=1.53351G/s null_percent=0.01 size=1048.58k

real    0m59.902s
user    0m59.880s
sys     0m0.020s
```

After: **1.2s** total run time
```
$ time release/arrow-compute-aggregate-benchmark --benchmark_filter="ModeKernelNarrow<BooleanType>/1048576/10000"
------------------------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------
ModeKernelNarrow<BooleanType>/1048576/10000        637 us          637 us         1098 bytes_per_second=1.53253G/s null_percent=0.01 size=1048.58k

real    0m1.210s
user    0m1.202s
sys     0m0.008s
```