@cyb70289 2021-05-10
Tested on Xeon Gold 5218.

Before: **5.4s** total run time
```
$ time release/arrow-compute-aggregate-benchmark --benchmark_filter="ModeKernelNarrow<BooleanType>/1048576/10000"
------------------------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------
ModeKernelNarrow<BooleanType>/1048576/10000        550 us          550 us         1279 bytes_per_second=1.7744G/s null_percent=0.01 size=1048.58k

real    0m5.445s
user    0m5.417s
sys     0m0.008s
```

After: **1.2s** total run time
```
$ time release/arrow-compute-aggregate-benchmark --benchmark_filter="ModeKernelNarrow<BooleanType>/1048576/10000"
------------------------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------
ModeKernelNarrow<BooleanType>/1048576/10000        551 us          551 us         1267 bytes_per_second=1.7723G/s null_percent=0.01 size=1048.58k

real    0m1.197s
user    0m1.180s
sys     0m0.000s
```