@cyb70289 2021-05-10
Timed running all arrow-compute-*-benchmark tests on x86, clang. Total time drops from 26 to 18 min.