@cyb70289 2021-05-10
Travis CI passed at https://travis-ci.org/github/cyb70289/arrow/builds/770391382